[Modify] 주소 공유 몽고디비에 저장처리

+++ src/main/java/com/example/meettify/controller/chat/ChatControllerDocs.java
@@ -9,9 +9,5 @@ import org.springframework.security.core.userdetails.UserDetails;
 @Tag(name = "채팅", description = "채팅 API")
 public interface ChatControllerDocs {
     @Operation(summary = "채팅", description = "채팅 생성하는 API")
-    void sendMessage(ChatMessageDTO message, int roomId);
-
-//    @Operation(summary = "채팅", description = "채팅 생성하는 API")
-//    void receive(ChatMessageDTO chatDTO);
-
+    void sendMessage(ChatMessageDTO message, Long roomId);
 }
